Fixed scaling problem with Bayesian priors.
Priors (parameters) can be either of the range 1-100 or ~ -1e-6 to 10e-6.
Priors or the smaller values were not being picked up, since the sum over priors was being
dominated by the larger values.

The solution is to scale all priors and parameters to the range between 0 to 1 using the
built in 'rescale.m' function (in gaussLogPrior.m)

Also, moved from using log(likelihood) to just likelihood. This is because likelihood is
exp(-chi^2/2), so log-likelihood is log(exp(x)). Exp x can be very small (approaching eps),
so log(exp... is numerically unwise.

Similarly, we use sum(priors) rather than log(sum(priors))